Overload ModelManager.java with Sochedule functions; Remove SocModelManager.java
