+++ src/main/java/seedu/address/model/ModelManager.java
@@ -23,7 +23,10 @@ public class ModelManager implements Model {
 
     private final AddressBook addressBook;
     private final UserPrefs userPrefs;
+    private final Sochedule sochedule;
     private final FilteredList<Person> filteredPersons;
+    private final FilteredList<Task> filteredTasks;
+    private final FilteredList<Event> filteredEvents;
 
 
     /**
@@ -38,18 +41,65 @@ public class ModelManager implements Model {
         this.addressBook = new AddressBook(addressBook);
         this.userPrefs = new UserPrefs(userPrefs);
         filteredPersons = new FilteredList<>(this.addressBook.getPersonList());
+
+        // initialised to null to maintain original functionality temporarily
+        this.sochedule = null;
+        filteredTasks = null;
+        filteredEvents = null;
     }
-
-
-    public ModelManager() {
-        this(new AddressBook(), new UserPrefs());
-    }
-
 
     /**
      * Initializes a ModelManager with the given sochedule and userPrefs.
      */
+    public ModelManager(ReadOnlySochedule sochedule, ReadOnlyUserPrefs userPrefs) {
+        super();
+        requireAllNonNull(sochedule, userPrefs);
 
+        logger.fine("Initializing with SOChedule: " + sochedule + " and user prefs " + userPrefs);
+
+        this.sochedule = new Sochedule(sochedule);
+        this.userPrefs = new UserPrefs(userPrefs);
+        filteredTasks = new FilteredList<>(this.sochedule.getTaskList());
+        filteredEvents = new FilteredList<>(this.sochedule.getEventList());
+
+        // initialised to null to maintain original functionality temporarily
+        this.addressBook = null;
+        filteredPersons = null;
+    }
+
+    // original function to be replaced once Sochedule support is complete
+    public ModelManager() {
+        this(new AddressBook(), new UserPrefs());
+    }
+
+    /*
+    public ModelManager() {
+        this(new Sochedule(), new UserPrefs());
+    }
+    */
+
+    @Override
+    public boolean equals(Object obj) {
+        // short circuit if same object
+        if (obj == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(obj instanceof ModelManager)) {
+            return false;
+        }
+
+        // state check
+        // null check is temporary during transition to new data
+        ModelManager other = (ModelManager) obj;
+        return (addressBook == null || addressBook.equals(other.addressBook)) //to be removed
+                && userPrefs.equals(other.userPrefs)
+                && (filteredPersons == null || filteredPersons.equals(other.filteredPersons)) //to be removed
+                && (sochedule == null || sochedule.equals(other.sochedule))
+                && (filteredTasks == null || filteredTasks.equals(other.filteredTasks))
+                && (filteredEvents == null || filteredEvents.equals(other.filteredEvents));
+    }
 
     //=========== UserPrefs ==================================================================================
 
@@ -87,6 +137,7 @@ public class ModelManager implements Model {
     }
 
     //=========== AddressBook ================================================================================
+    // Functions of the original AB3 - will be deleted once transition over to new data structure
 
     @Override
     public void setAddressBook(ReadOnlyAddressBook addressBook) {
@@ -123,7 +174,7 @@ public class ModelManager implements Model {
     }
 
     //=========== Filtered Person List Accessors =============================================================
-
+    // Functions of the original AB3 - will be deleted once transition over to new data structure
     /**
      * Returns an unmodifiable view of the list of {@code Person} backed by the internal list of
      * {@code versionedAddressBook}
@@ -139,134 +190,123 @@ public class ModelManager implements Model {
         filteredPersons.setPredicate(predicate);
     }
 
-    @Override
-    public boolean equals(Object obj) {
-        // short circuit if same object
-        if (obj == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(obj instanceof ModelManager)) {
-            return false;
-        }
-
-        // state check
-        ModelManager other = (ModelManager) obj;
-        return addressBook.equals(other.addressBook)
-                && userPrefs.equals(other.userPrefs)
-                && filteredPersons.equals(other.filteredPersons);
-    }
-
     //=========== task ==================================================================================
 
     @Override
     public Path getTaskListFilePath() {
-        //THIS IS A DUMMY RETURN
-        return null;
+        return userPrefs.getTaskListFilePath();
     }
 
-
     @Override
-    public void setTaskListFilePath(Path taskListFilePath){
+    public void setTaskListFilePath(Path taskListFilePath) {
+        requireAllNonNull(taskListFilePath);
+        userPrefs.setTaskListFilePath(taskListFilePath);
     }
 
     @Override
     public void setTaskList(ReadOnlySochedule sochedule) {
-
+        this.sochedule.resetTaskData(sochedule);
     }
 
     @Override
     public ReadOnlySochedule getTaskList() {
-        return null;
+        return sochedule;
     }
 
     @Override
     public boolean hasTask(Task task) {
-        //THIS IS A DUMMY METHOD
-        return true;
+        requireAllNonNull(task);
+        return sochedule.hasTask(task);
     }
 
     @Override
     public void deleteTask(Task target) {
-
+        sochedule.removeTask(target);
     }
 
     @Override
     public void addTask(Task task) {
-
+        sochedule.addTask(task);
+        updateFilteredTaskList(PREDICATE_SHOW_ALL_TASKS);
     }
 
     @Override
     public void setTask(Task target, Task editedTask) {
-
+        requireAllNonNull(target, editedTask);
+        sochedule.setTask(target, editedTask);
     }
 
+    /**
+     * Returns an unmodifiable view of the list of {@code Task}
+     */
     @Override
     public ObservableList<Task> getFilteredTaskList() {
-        //THIS IS A DUMMY METHOD
-        return null;
+        return filteredTasks;
     }
 
     @Override
     public void updateFilteredTaskList(Predicate<Task> predicate) {
-
+        requireAllNonNull(predicate);
+        filteredTasks.setPredicate(predicate);
     }
 
     //=========== event ==================================================================================
 
     @Override
     public Path getEventListFilePath() {
-        //THIS IS A DUMMY RETURN
-        return null;
+        return userPrefs.getTaskListFilePath();
     }
 
     @Override
-    public void setEventListFilePath(Path eventListFilePath){
-
+    public void setEventListFilePath(Path eventListFilePath) {
+        requireAllNonNull(eventListFilePath);
+        userPrefs.setEventListFilePath(eventListFilePath);
     }
 
     @Override
     public void setEventList(ReadOnlySochedule sochedule) {
-
+        this.sochedule.resetEventData(sochedule);
     }
 
     @Override
     public ReadOnlySochedule getEventList() {
-        //THIS IS A DUMMY RETURN
-        return null;
+        return sochedule;
     }
 
     @Override
     public boolean hasEvent(Event event) {
-        //THIS IS A DUMMY METHOD
-        return true;
+        requireAllNonNull(event);
+        return sochedule.hasEvent(event);
     }
 
     @Override
     public void deleteEvent(Event target) {
-
+        sochedule.removeEvent(target);
     }
 
     @Override
     public void addEvent(Event event) {
-
+        sochedule.addEvent(event);
+        updateFilteredEventList(PREDICATE_SHOW_ALL_EVENTS);
     }
 
     @Override
     public void setEvent(Event target, Event editedEvent) {
-
+        requireAllNonNull(target, editedEvent);
+        sochedule.setEvent(target, editedEvent);
     }
 
+    /**
+     * Returns an unmodifiable view of the list of {@code Event}
+     */
     @Override
     public ObservableList<Event> getFilteredEventList() {
-        //THIS IS A DUMMY METHOD
-        return null;
+        return filteredEvents;
     }
 
     @Override
     public void updateFilteredEventList(Predicate<Event> predicate) {
-
+        requireAllNonNull(predicate);
+        filteredEvents.setPredicate(predicate);
     }
-
 }
